% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{counts2logROR}
\alias{counts2logROR}
\title{obtain logR and logOR from read counts and GC-correct logR}
\usage{
counts2logROR(
  mat,
  gbuild,
  unmatched = FALSE,
  MandUnormal = FALSE,
  f,
  spanT,
  spanA,
  spanX
)
}
\description{
obtain logR and logOR from read counts and GC-correct logR
}
