% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets2n-package.R
\docType{package}
\name{facets2n-package}
\alias{facets2n-package}
\title{package documentation}
\description{
package documentation
}
\details{
{FACETS stands for Fraction and Allelic Copy number Estimation from Tumor/normal Sequencing. It uses high coverage next generation sequencing data (whole exome or targeted) to estimate the total and integer allele specific copy numbers as well as the tumor purity and ploidy. The methodology is described in the reference paper. This package created to test FACETS with both matched and unmatched normal for log ratio normalization.}
}
\references{
{Shen, R and Seshan, VE (2016). FACETS: allele-specific copy number and clonal heterogeneity analysis tool for high-throughput DNA sequencing.\emph{Nucleic Acids Research} \url{https://doi.org/10.1093/nar/gkw520}}
}
