% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{FindBestNormalParameters}
\alias{FindBestNormalParameters}
\title{FindBestNormalParameters takes takes a facets2n generated tumor loess object and snp-pileup generated pileup file, and optional similar files for reference normals, and returns the pileup data for the best normal for T/N CNLR.}
\usage{
FindBestNormalParameters(
  TumorLoess,
  TumorPileup,
  ReferenceLoess = NULL,
  ReferencePileup = NULL,
  MinOverlap = 0.9,
  useMatchedX = FALSE
)
}
\arguments{
\item{TumorLoess}{(matrix) A facets2n generated TumorLoess matrix with header and span values in the first row.}

\item{TumorPileup}{(data frame) snp-pileup generated pileup data frame with sample columns that match with the TumorLoess object.}

\item{ReferenceLoess}{(matrix) A ReferenceLoess matrix with a header and span values in the first row.}

\item{MinOverlap}{(numeric) A numeric between 0 and 1 that denotes the fraction overlap of loci between TumorLoess and the optional ReferenceLoess}

\item{useMatchedX}{(logical) Force select matched normal for normalization in ChrX.}

\item{ReferencePileupFile}{(data frame) A snp-pileup generated pileup data frame with sample columns that match with the ReferenceLoess object.}
}
\value{
A list of data frame with pileup depth values of Tumor, matched Normal, and a best unmatched normal, and the associated span values.
}
\description{
FindBestNormalParameters takes takes a facets2n generated tumor loess object and snp-pileup generated pileup file, and optional similar files for reference normals, and returns the pileup data for the best normal for T/N CNLR.
}
