% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-wrapper.R
\name{logRlogORspider}
\alias{logRlogORspider}
\title{logRlogRsplide plot generation from cncf input}
\usage{
logRlogORspider(cncf, dipLogR = 0, nfrac = 0.005)
}
\description{
logRlogRsplide plot generation from cncf input
}
