% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{MakeLoessObject}
\alias{MakeLoessObject}
\title{MakeLoessObject takes a pipleup file generated by snp-pileup and generates a loess/lowess object, which is also optinally written into an output file.}
\usage{
MakeLoessObject(
  pileup,
  write.loess = FALSE,
  outfilepath = "./loess.txt",
  is.Reference = FALSE,
  gbuild = "hg19"
)
}
\arguments{
\item{pileup}{(data frame) A data franme of snp-pileup generated depth.}

\item{write.loess}{(logical) Write loess object into file, instead of returning it as a matrix?}

\item{outfilepath}{(character) Filepath for writing loess object.}

\item{is.Reference}{(logical) Indicate whether the snp-pilep is Reference data.}

\item{gbuild}{(character) genome build version.}
}
\value{
A dataframe of loess normalized values for all input samples against filtered loci or None, if the loess normalized value is to be written to an output file.
}
\description{
MakeLoessObject takes a pipleup file generated by snp-pileup and generates a loess/lowess object, which is also optinally written into an output file.
}
