% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-wrapper.R
\name{plotSample}
\alias{plotSample}
\title{Plot the data and results for a single sample}
\usage{
plotSample(
  x,
  emfit = NULL,
  clustered = FALSE,
  plot.type = c("em", "naive", "both", "none"),
  sname = NULL
)
}
\description{
Plots copy number log-ratio, variant allele log-odds ratio as well as the copy number and cellular fraction fits.
}
