% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{PreProcSnpPileup}
\alias{PreProcSnpPileup}
\title{PreProcSnpPileup takes a snp-pileup generated pileup file and process it into a data frame format for downstream processing.}
\usage{
PreProcSnpPileup(
  filename,
  err.thresh = Inf,
  del.thresh = Inf,
  is.Reference = FALSE,
  gbuild = "hg19"
)
}
\arguments{
\item{filename}{(character) File path to snp-pileup generated pileup file.}

\item{err.thresh}{(numeric) Error threshold to be used to filter snp-pileup data frame.}

\item{del.thresh}{(numeric) Deletion threshold to be used to filter snp-pileup data frame.}

\item{is.Reference}{(logical) Indicate whether the snp-pilep is Reference data.}

\item{gbuild}{(character) genome build version.}
}
\value{
A data drame of pileup depth values filtered against del.thresh and err.thresh values.
}
\description{
PreProcSnpPileup takes a snp-pileup generated pileup file and process it into a data frame format for downstream processing.
}
