% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-wrapper.R
\name{procSample}
\alias{procSample}
\title{Process a sample}
\usage{
procSample(x, cval = 150, min.nhet = 15, dipLogR = NULL)
}
\arguments{
\item{x}{the output from preProcSample; should contain seg.tree and jointseg}

\item{cval}{critical value for segmentation}

\item{min.nhet}{minimum number of heterozygote snps in a segment used for bivariate t-statistic during clustering of segments}

\item{dipLogR}{diploid level obtained from a fit, typically using a higher cval, can be used with lower cval to recover focal changes}
}
\value{
\item{jointseg}{The data that were segmented. Only the loci that were sampled within a snp.nbhd are present}
\item{out}{data frame of segment summaries pre and post clustering of segments. The columns are: chrom the chromosome to which the segment belongs; seg the segment number; num.mark the number of SNPs in the segment; nhet the number of SNPs that are deemed heterozygous; cnlr.median the median log-ratio of the segment; mafR the log-odds-ratio summary for the segment; segclust the segment cluster to which segment belongs; cnlr.median.clust the median log-ratio of the segment cluster; mafR.clust the log-odds-ratio summary for the segment cluster; cf the cellular fraction of the segment; tcn the total copy number of the segment; lcn the minor copy number of the segment.}
\item{dipLogR}{specified or obtained from data}
\item{...}{other output when findDiploidLogR is used}
}
\description{
Processes the output from preProcSample for given cval and min.nhet
}
\details{
The minor copy number lcn may not be estimated with confidence when a segment has fewer than min.nhet heterozygous SNPs and hence will return NA. If there are too few heterozygous SNPs in a segment then mafR and mafR.clust can be NA.
}
