% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{procSnps}
\alias{procSnps}
\title{heterozygous and keep flags of the SNPs}
\usage{
procSnps(
  rcmat,
  ndepth = 35,
  het.thresh = 0.25,
  snp.nbhd = 250,
  gbuild = "hg19",
  unmatched = FALSE,
  ndepthmax = 5000
)
}
\description{
heterozygous and keep flags of the SNPs
}
