% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{procXSnps}
\alias{procXSnps}
\title{Takes a snp-pileup file and determines sex of sample and any unmatched normals based on number of chrX het SNPs, as males should not have het X}
\usage{
procXSnps(
  pileup,
  ndepth = 35,
  het.thresh = 0.25,
  snp.nbhd = 250,
  gbuild = "hg19",
  unmatched = FALSE,
  ndepthmax = 5000,
  nhet = 10
)
}
\arguments{
\item{pileup}{(character) A snp-pileup generated pileup file that has been analyzed with readSnpMatrix(). Expect columns "Chromosome", "Position", "NOR.DP", "NOR.RD", "TUM.DP", and "TUM.RD". Can be a pileup file that has been merged with common loci of reference normals processed with PreProcSnpPileup()}

\item{ndepth}{(numeric) minimum normal sample depth to keep}

\item{het.thresh}{(numeric) vaf threshold to call a SNP heterozygous}

\item{snp.nbhd}{(logical) window size}

\item{gbuild}{(character) genome build version.}

\item{unmatched}{(logical)}

\item{ndepthmax}{(numeric) loci for which normal coverage exceeds this number (default is 5000) will be discarded as PCR duplicates. Fof high coverage sample increase this and ndepth commensurately.}

\item{nhet}{(numeric) minimum number of heterzygous SNPs to classify sample as Female}
}
\value{
output a table of samples analyzed and imputed sex.
}
\description{
Takes a snp-pileup file and determines sex of sample and any unmatched normals based on number of chrX het SNPs, as males should not have het X
}
