% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{scanSnp}
\alias{scanSnp}
\title{scan maploc for snps that are close to one another (within snp.nbhd bases)
keep all the hets (should change if too close) and only one from a nbhd}
\usage{
scanSnp(maploc, het, nbhd)
}
\arguments{
\item{maploc}{(numeric) chrom genomic mapping position}

\item{het}{(numeric) binary heterozygous status of position}

\item{nbhd}{(numeric) window size, bp}
}
\value{
a vector of snp positions to keep
}
\description{
scan maploc for snps that are close to one another (within snp.nbhd bases)
keep all the hets (should change if too close) and only one from a nbhd
}
