% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{segmentCBS}
\alias{segmentCBS}
\title{segmentCBS takes cnlr values from all snps, segments them with CBS and the clusters the segments. Probes with distribution closest to clnr of 0 are used as null distribution for assigning p-values to segments}
\usage{
segmentCBS(seglist, make.plots = TRUE, sname = NULL, outdir = NULL)
}
\arguments{
\item{seglist}{(list) list returned from preProcSample}

\item{make.plots}{(logical) make segmentation plots}

\item{sname}{(character) sample ID}

\item{outdir}{(character) directory for plotting output}
}
\description{
segmentCBS takes cnlr values from all snps, segments them with CBS and the clusters the segments. Probes with distribution closest to clnr of 0 are used as null distribution for assigning p-values to segments
}
