% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{segmentTransplant}
\alias{segmentTransplant}
\title{segmentTransplant takes cnlr values from all snps, segments them with CBS and the clusters the segments. Probes with distribution closest to clnr of 0 are used as null distribution for assigning p-values to segments}
\usage{
segmentTransplant(seglist, make.plots = TRUE, sname = NULL, outdir = NULL)
}
\arguments{
\item{seglist}{(list) list returned from preProcSample}

\item{make.plots}{(logical) make segmentation plots}

\item{sname}{(character) sample ID}

\item{outdir}{(character) directory for plotting output}
}
\description{
segmentTransplant takes cnlr values from all snps, segments them with CBS and the clusters the segments. Probes with distribution closest to clnr of 0 are used as null distribution for assigning p-values to segments
}
