% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets-procreads.R
\name{procSnps}
\alias{procSnps}
\title{heterozygous and keep flags of the SNPs}
\usage{
procSnps(
  rcmat,
  ndepth = 35,
  het.thresh = 0.25,
  snp.nbhd = 250,
  gbuild = "hg19",
  unmatched = FALSE,
  ndepthmax = 5000,
  donorCounts = NULL
)
}
\arguments{
\item{rcmat}{input counts matrix}

\item{ndepth}{(numeric) minimum normal sample depth to keep}

\item{het.thresh}{(numeric) vaf threshold to call a SNP heterozygous}

\item{snp.nbhd}{(logical) window size}

\item{gbuild}{(character) genome build version.}

\item{unmatched}{(logical)}

\item{ndepthmax}{(numeric) loci for which normal coverage exceeds this number (default is 5000) will be discarded as PCR duplicates. Fof high coverage sample increase this and ndepth commensurately.}

\item{donorCounts}{count matrix of donor sample(s)}
}
\description{
heterozygous and keep flags of the SNPs
}
